refactor(indexer/rpc): extract breaker fetchers + probeFunction + cache to rpc/breakers.ts (PR-S9)
Move ~487 lines of breaker-related code out of rpc.ts (642 → 154)
into the new indexer-envio/src/rpc/breakers.ts (513 lines):
- BreakerKindRpc, BreakerDefaults, BreakerFeedState types
- _testBreaker* mocks (kinds, defaults, feed-state, list) + setters/clearers
- breakerKindKey, breakerFeedStateKey, parseRateFeedBreakerStatus helpers
- probeFunction
- fetchBreakerList, fetchBreakerKind, fetchBreakerDefaults, fetchBreakerFeedState

Cross-chain isolation via chainId-keyed cache state is preserved
verbatim. The Oracle DB query helpers (getPoolsByFeed,
updatePoolsOracleExpiry, getPoolsWithReferenceFeed,
getBreakerConfigsByFeed) stay in rpc.ts alongside the barrel
re-exports — DB-backed, not RPC-backed.

After this slice, rpc.ts is 154 lines (barrel + Oracle DB helpers)
and the four rpc/* modules each carry a single responsibility.

rpc.ts re-exports the previously-public surface so handlers
(breakers.ts, EventHandlers.ts) and tests keep working without
import-path changes.

Pure extraction. 380 tests pass.
